requirements.txt
openpyxl
scraperwiki_local
lxml
